requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic-settings>=2.7.0
livekit-agents~=1.4
livekit-plugins-elevenlabs~=1.1
livekit-plugins-silero~=1.1
livekit-plugins-google~=1.4
livekit-api>=1.0.0
structlog>=24.4.0
python-dotenv>=1.0.0
ruff>=0.8.0
